Version 3.0 BETA

+++ www/admin/hook/widgetDelete.html.php
@@ -4,8 +4,10 @@
 Параметры: string $data[0] - имя виджета, int $data[1] - идентификатор виджета, mixed $data[2] - параметры виджета */
 
 if($data[0]!='html') return true;
-$f=core::path().'data/widgetHtml/'.$data[2].'.html';
-if(!file_exists($f)) return true;
-unlink($f);
+$cfg=core::config();
+foreach($cfg['languageList'] as $item) {
+	$f=core::path().'data/widgetHtml/'.$data[2].'_'.$item.'.html';
+	if(file_exists($f)) unlink($f);
+}
 return true;
 ?>
